feat: added Text and AVG Fps tracker

+++ src/class/IAppProcess.h
@@ -5,7 +5,7 @@
 class IAppProcess {
 public:
     virtual bool init() = 0;
-    virtual void update(float* dt) = 0;
+    virtual void update(double* dt) = 0;
     virtual void render(SDL_Renderer* renderer) = 0;
     virtual void close() = 0;
 };
